♻️ move stored data
