+++ src/components/data.tsx
@@ -24,3 +24,5 @@ const Stations: Station[] = [
 ];
 
 export default Stations;
+
+export const VERSION = "1.0";
